Changed Quantity variable from int to double

+++ src/main/java/seedu/duke/command/misc/AddCommand.java
@@ -19,14 +19,14 @@ public class AddCommand extends ExecutableCommand {
         this.amount = flag;
     }
 
-    private static void addToExistingIngredients(DukeSession dukeSession, int quantity, int index) {
-        int newQuantity = DukeSession.ingredients.get(index).getQuantity() + quantity;
+    private static void addToExistingIngredients(DukeSession dukeSession, Double quantity, int index) {
+        double newQuantity = DukeSession.ingredients.get(index).getQuantity() + quantity;
         DukeSession.ingredients.get(index).setQuantity(newQuantity);
         dukeSession.getUi().printMessage("Here is the new quantity of the ingredient:");
         dukeSession.getUi().printMessage(String.valueOf(DukeSession.ingredients.get(index)));
     }
 
-    private static void addNewIngredient(DukeSession dukeSession, int quantity, String name) {
+    private static void addNewIngredient(DukeSession dukeSession, Double quantity, String name) {
         Ingredient ingredient = new Ingredient(name, quantity);
         DukeSession.ingredients.add(ingredient);
         dukeSession.getUi().printMessage("the following ingredient has been added");
@@ -44,7 +44,7 @@ public class AddCommand extends ExecutableCommand {
 
     public void execute(DukeSession dukeSession) {
         try {
-            int quantity = Integer.parseInt(amount);
+            Double quantity = Double.parseDouble(amount);
             if (quantity <= 0) {
                 throw new DukeException("OOPS, quantity must be greater than 0");
             }
